WIP reverse typing
